Deleted mood and patient and added more accurate database model 'appUser'
Added appUser to better serve upcoming database models
