+++ backend/app.js
@@ -1,8 +1,6 @@
 const middleware = require('./utils/middleware')
-//const patientsRouter = require('./controllers/patientsRouter')
-const patientsRouter = require('./routes/patientRouter')
+const userRouter = require('./routes/userRouter')
 const express = require('express')
-// Try-catch ei tarvita patientsRouterissa
 require('express-async-errors')
 const morgan = require('morgan')
 const cors = require('cors')
@@ -12,10 +10,9 @@ const app = express()
 // Tama tarvitaan, jos frontend buildataan backendiin
 //app.use(express.static('build'))
 
-// JSON luettavaan muotoon
 app.use(express.json())
 
-// Morgan logit HTTP-pyynnöistä
+// Morgan logs for HTTP-queries
 morgan.token('body', (req) => JSON.stringify(req.body))
 app.use(morgan('tiny'))
 app.use(morgan(':body'))
@@ -23,15 +20,20 @@ app.use(morgan(':body'))
 app.use(cors())
 
 const db = require('./models')
+// use alter: true if you need to change database models without losing test/dummy data
+// db.sequelize.sync({ alter: true })
+
+
 db.sequelize.sync()
 
-// Tietokannan url
-const patientsUrl = '/api'
-// /api/patients/ polkujen käsittely
-app.use(patientsUrl, patientsRouter)
-// Pyyntoihin ylla maarittelemattomista poluista palautetaan 404
+// Database url
+const usersUrl = '/api'
+app.use(usersUrl, userRouter)
+
+// Return 404 for undefined paths
 app.use(middleware.unknownEndpoint)
-// Virheidenkasittely
+
+// Error handling
 app.use(middleware.errorHandler)
 
 module.exports = app
